requirements.txt
Flask
boto3
mysql-connector-python
gunicorn